# swagger.yaml
openapi: 3.0.0
info:
  title: Kiwibot delivery service
  description: A Kiwibot API for managing bots and deliveries.
  contact:
    name: Angela Cobos
    email: "cobos.haydee@gmail.com"
  version: "1.0"
tags:
  - name: Bots
    description: Endpoints related to managing Kiwibot's automated bots.
  - name: Deliveries
    description: Endpoints related to deliveries.
paths:
  /bots:
    post:
      summary: Creates a new bot
      tags: [Bots]
      responses:
        200:
          description: Bot created succesfully.
  /bots/{zone}:
    get:
      summary: Get all bots in a zone.
      tags: [Bots]
      parameters:
        - in: path
          name: zone
          schema:
            type: string
          required: true
      responses:
        200:
          description: All bots in the zone.
        500:
          description: Some server error.
  /bots/{zone}/assign:
    post:
      summary: Assigns all available bots in a zone to pending deliveries.
      tags: [Bots]
      parameters:
        - in: path
          name: zone
          schema:
            type: string
          required: true
      responses:
        200:
          description: Bots assigned successfully.
        500:
          description: Some server error.
  /deliveries:
    post:
      summary: Creates a new delivery.
      tags: [Deliveries]
      responses:
        200:
          description: Delivery created successfully.
    get:
      summary: Get all deliveries, allows for pagination.
      tags: [Deliveries]
      parameters:
        - in: query
          name: offset
          schema:
            type: number
          required: false
        - in: query
          name: limit
          schema:
            type: number
          required: false
      responses:
        200:
          description: All deliveries.
        500:
          description: Some server error.
  /deliveries/{id}:
    get:
      summary: Gets a single delivery by id.
      tags: [Deliveries]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        200:
          description: The requested delivery.
        404:
          description: No delivery with that id was found.
        500:
          description: Some server error.